+++ lib/slop/options.rb
@@ -1,18 +1,6 @@
 class Slop
   class Options < Array
 
-    # @param [Boolean] symbols true to cast hash keys to symbols
-    # @see Slop#to_hash
-    # @return [Hash]
-    def to_hash(symbols)
-      reduce({}) do |hsh, option|
-        key = option.key
-        key = key.to_sym if symbols
-        hsh[key] = option.argument_value
-        hsh
-      end
-    end
-
     # Fetch an Option object. This method overrides Array#[] to provide
     # a nicer interface for fetching options via their short or long flag.
     # The reason we don't use a Hash here is because an option cannot be