+++ CHANGES.md
@@ -1,6 +1,10 @@
-1.10.0 (TBA)
-------------
+2.0.0 (TBA)
+-----------
 
+* Deprecations:
+  * Removed `Slop::Options#to_hash` continue using `Slop#to_hash` directly.
+    This method also now returns symbols by default instead of strings. If
+    you want strings use `opts.to_hash(false)`
 * Added Slop::Error subclassing from StandardError which all exception
   classes should inherit from
 * Added Slop::MissingOptionError and `:required` option to Slop::Option.